Made single chrome api module. Reason is that crome api is tightly coupled. And I have no desire to fight with it. Btw external declaration have low impact on compilation result. So it's ok

+++ component/chrome/src/commonMain/kotlin/chrome/bookmarks/BookmarkTreeNode.kt
@@ -35,7 +35,7 @@ external interface BookmarkTreeNode {
 	var id: String
 
 	/** The 0-based position of this node within its parent folder. */
-	var index: Double?
+	var index: Int?
 
 	/** The `id` of the parent folder. Omitted for the root node. */
 	var parentId: String?
